Rename blocks/shared => blocks/types
