+++ addons/docs/src/frameworks/react/extractProps.ts
@@ -2,7 +2,7 @@ import PropTypes from 'prop-types';
 import { isForwardRef, isMemo } from 'react-is';
 import { PropDef } from '@storybook/components';
 import { hasDocgen, extractComponentProps, PropsExtractor, TypeSystem } from '../../lib/docgen';
-import { Component } from '../../blocks/shared';
+import { Component } from '../../blocks/types';
 import { enhancePropTypesProps } from './propTypes/handleProp';
 import { enhanceTypeScriptProps } from './typeScript/handleProp';
